+++ src/__init__.py
@@ -1,0 +1,13 @@
+from pathlib import Path
+
+# Solar Zenith Angle
+LATITUDE = 50.110656
+LONGITUDE = 8.682526
+ALTITUDE = 350_000
+
+# Data
+DATA_IN = Path("..", "data", "in")
+
+# MLFlow
+ML_SERVER_URI = "http://localhost:5000"
+EXPERIMENT_NAME = "CatBoost"
